Fixed deserialization objct casting bug

+++ src/main/java/net/whg/utils/events/location/CylinderLocationTrigger.java
@@ -26,8 +26,8 @@ public record CylinderLocationTrigger(String name, Location center, float radius
         var x = (double) args.get("x");
         var y = (double) args.get("y");
         var z = (double) args.get("z");
-        var r = (float) args.get("r");
-        var h = (float) args.get("h");
+        var r = (float) (double) args.get("r");
+        var h = (float) (double) args.get("h");
         var triggerOnTeleport = (boolean) args.get("triggerOnTeleport");
 
         var center = new Location(world, x, y, z);
